Created a the productsPage, the page shows the chosen product, information regarding the product, a review section and actions to add the product to the cart.

+++ BE/src/models/review.model.ts
@@ -45,7 +45,7 @@ Review.init(
             allowNull: false,
         },
         starReview: {
-            type: DataTypes.DECIMAL(1),
+            type: DataTypes.INTEGER,
             allowNull: false,
             validate: {
                 min: 1,
